[#199]context api 따로 분리
