+++ src/pages/Visualization/components/RightSection/RightSection.tsx
@@ -1,5 +1,6 @@
 import { useState, useContext, useEffect, useRef, useCallback } from "react";
-import { CodeContext, PreprocessedCodesContext } from "../../Visualization";
+import { CodeContext } from "../../context/CodeContext";
+import { PreprocessedCodesContext } from "../../context/PreProcessedCodesContext";
 import Split from "react-split";
 import _ from "lodash";
 import ResizeObserver from "resize-observer-polyfill";
