+++ src/pages/Visualization/context/PreProcessedCodesContext.ts
@@ -1,0 +1,13 @@
+import { createContext, SetStateAction, Dispatch } from "react";
+import { ValidTypeDto } from "../types/dto/ValidTypeDto";
+
+// 전처리한 코드 타입 정의
+interface PreprocessedCodeContextType {
+  preprocessedCodes: ValidTypeDto[];
+  setPreprocessedCodes: Dispatch<SetStateAction<ValidTypeDto[]>>;
+}
+
+export const PreprocessedCodesContext = createContext<PreprocessedCodeContextType>({
+  preprocessedCodes: [],
+  setPreprocessedCodes: () => {},
+});
